+++ src/main/java/interface_adapter/pokemon_lookup/PokemonLookupViewModel.java
@@ -1,4 +1,4 @@
-package interface_adapter.pokemon;
+package interface_adapter.pokemon_lookup;
 
 import interface_adapter.ViewModel;
